Fix selection issues when drums are absent
If multiple cannons were placed into a scene w/o drums, attempting to select any cannon but the first would fail. This happened because the code used 1 loop to check if it should select a cannon and if it should select a drum, causing an out of bounds error if the iteration passed the end of the drum list before finding the appropriate cannon.